[.cleanup_gmx_analyses.sh]: Refactor

+++ analysis/lintf2_ether/gmx/.cleanup_gmx_analyses.sh
@@ -8,31 +8,29 @@
 gather() {
     # Gather output of similar analyses scripts.
     root="${1}"
-    mkdir "${root}" || exit
-    mv "${root}"_*slurm-[0-9]* "${root}"/
-    cd "${root}"/ || exit
-    mv "${root}"_*slurm-[0-9]*/* ./
-    rm -r "${root}"_*slurm-[0-9]*
-    cd ../ || exit
+    if [[ -d ${root} ]]; then
+        echo "WARNING: Directory already exists: '${root}'"
+        return 0
+    fi
+    if [[ -n $(find . -maxdepth 1 -type d -name "${root}_*slurm-[0-9]*" -print -quit) ]]; then
+        mkdir "${root}/" || exit
+        mv "${root}"_*slurm-[0-9]*/ "${root}/" || exit
+        cd "${root}/" || exit
+        mv "${root}"_*slurm-[0-9]*/* ./ || exit
+        rm -r "${root:?}"_*slurm-[0-9]*/ || exit
+        cd ../ || exit
+    else
+        echo "WARNING: No directory matching pattern '${root}_*slurm-[0-9]*'"
+    fi
 }
 
-mv energy_slurm-[0-9]* energy
-mv make_ndx_slurm-[0-9]* make_ndx
-mv polystat_slurm-[0-9]* polystat
-mv potential-z_slurm-[0-9]* potential-z
-
-gather density-z
-gather densmap-z
-gather msd
-gather trjconv
-
-name1=rdf
-name2=rdf_slab-z
-gather "${name2}"
-mkdir "${name1}" || exit
-mv "${name1}"_* "${name1}"/
-cd "${name1}"/ || exit
-mv "${name2}"/ ../
-mv "${name1}"_*/* ./
-rm -r "${name1}"_*
-cd ../ || exit
+gather "density-z"
+gather "densmap-z"
+gather "energy"
+gather "make_ndx"
+gather "msd"
+gather "polystat"
+gather "potential-z"
+gather "rdf_slab-z"
+gather "rdf"
+gather "trjconv"
